requirements.txt
# Core dependencies
flask>=3.0
gunicorn>=21.2.0
pycups>=2.0.4
python-magic>=0.4.27
python-dotenv>=1.0

# Database
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
psycopg2-binary>=2.9

# Authentication
Authlib>=1.3
httpx>=0.27
requests>=2.31

# Utilities
shortuuid>=1.0
pypdf>=4.0

# Development dependencies
pytest>=8.0
pytest-cov>=4.0
